@@ -1,77 +1,67 @@
 {
   "name": "trainee-ui-app",
-  "version": "0.55.6",
+  "version": "0.55.8",
   "private": true,
   "dependencies": {
-    "@aws-amplify/ui-react": "^2.20.0",
-    "@cypress/code-coverage": "^3.9.12",
+    "@aws-amplify/ui-react": "^3.6.0",
+    "@cypress/code-coverage": "^3.10.0",
     "@cypress/instrument-cra": "^1.4.0",
-    "@cypress/react": "^5.8.0",
-    "@cypress/webpack-dev-server": "^1.4.0",
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@fortawesome/fontawesome-common-types": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-regular-svg-icons": "^6.0.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@mui/material": "^5.8.7",
-    "@reduxjs/toolkit": "^1.7.1",
-    "@sentry/browser": "^6.8.0",
-    "@types/node": "^14.14.32",
-    "@types/qrcode.react": "^1.0.1",
-    "@types/react": "^17.0.5",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-gtm-module": "^2.0.0",
-    "@types/react-helmet": "^6.1.4",
-    "@types/react-router-dom": "^5.1.5",
-    "@types/yup": "0.29.13",
-    "aws-amplify": "^4.3.25",
-    "axios": "^0.21.2",
-    "babel-plugin-istanbul": "^6.0.0",
-    "browser-update": "^3.3.38",
-    "cross-env": "^7.0.3",
-    "cypress": "^8.7.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@reduxjs/toolkit": "^1.9.0",
+    "@sentry/browser": "^7.21.1",
+    "@types/cypress__code-coverage": "^3.10.0",
+    "@types/jest": "^29.1.2",
+    "@types/node": "^18.11.09",
+    "@types/qrcode.react": "^1.0.2",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-gtm-module": "^2.0.1",
+    "@types/react-helmet": "^6.1.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-test-renderer": "^17.0.1",
+    "aws-amplify": "^4.3.43",
+    "axios": "^1.2.0",
+    "browser-update": "^3.3.44",
+    "cypress": "^11.2.0",
     "cypress-otp": "^1.0.3",
     "dayjs": "^1.10.7",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-cypress": "^2.11.3",
-    "eslint-plugin-prettier": "^3.4.0",
-    "formik": "^2.2.6",
-    "html-webpack-plugin": "^4.5.2",
-    "istanbul-lib-coverage": "^3.0.1",
-    "mapbox-gl": "^2.9.1",
-    "material-ui-confirm": "^3.0.4",
-    "nanoid": "^3.2.0",
-    "nhsuk-frontend": "^6.1.1",
+    "enzyme": "^3.11.0",
+    "eslint": "^8.28.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "formik": "^2.2.9",
+    "husky": "^8.0.2",
+    "jest-sonar-reporter": "^2.0.0",
+    "material-ui-confirm": "^3.0.7",
+    "mocha": "^10.1.0",
+    "mochawesome": "^7.1.3",
+    "mochawesome-merge": "^4.2.1",
+    "nanoid": "^4.0.0",
+    "nhsuk-frontend": "^6.1.2",
     "nhsuk-react-components": "^2.2.0",
-    "nyc": "^15.1.0",
-    "prettier": "^2.3.0",
-    "qrcode.react": "^3.0.2",
-    "react": "^17.0.2",
-    "react-cookie-consent": "^6.2.4",
-    "react-dom": "^17.0.2",
+    "prettier": "^2.8.0",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-gtm-module": "^2.0.11",
     "react-helmet": "^6.1.0",
     "react-loading": "^2.0.3",
-    "react-phone-number-input": "^3.1.41",
-    "react-redux": "^7.2.3",
+    "react-phone-number-input": "^3.2.12",
+    "react-redux": "^8.0.5",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-test-renderer": "^17.0.2",
-    "redux": "^4.0.5",
+    "react-scripts": "5.0.0",
+    "react-select": "^5.7.0",
+    "react-test-renderer": "^18.2.0",
+    "redux": "^4.2.0",
     "redux-mock-store": "^1.5.4",
     "redux-thunk": "^2.3.0",
     "sass": "^1.32.13",
-    "source-map-support": "^0.5.21",
-    "stylelint": "^13.12.0",
-    "ts-node": "^10.7.0",
-    "tsify": "^5.0.4",
-    "typescript": "^4.2.4",
-    "webpack-dev-server": "3.11.1",
+    "stylelint": "^14.15.0",
+    "stylelint-prettier": "^2.0.0",
+    "typescript": "^4.6.4",
     "yup": "^0.32.11"
   },
   "scripts": {
