@@ -32,10 +32,11 @@ const View = ({ canEdit, history }: IView) => {
             </p>
           </WarningCallout>
         )}
-        {FormRUtilities.displaySubmissionDate(
-          formData.submissionDate,
-          "submissionDateTop"
-        )}
+        {!canEdit &&
+          FormRUtilities.displaySubmissionDate(
+            formData.submissionDate,
+            "submissionDateTop"
+          )}
         <Card feature>
           <Card.Content>
             <Card.Heading>Personal Details</Card.Heading>
@@ -210,10 +211,11 @@ const View = ({ canEdit, history }: IView) => {
             </SummaryList>
           </Card.Content>
         </Card>
-        {FormRUtilities.displaySubmissionDate(
-          formData.submissionDate,
-          "submissionDate"
-        )}
+        {!canEdit &&
+          FormRUtilities.displaySubmissionDate(
+            formData.submissionDate,
+            "submissionDate"
+          )}
       </>
     );
   else content = <Redirect to="/formr-a" />;
