@@ -71,10 +71,11 @@ const View = ({ canEdit, history }: IView) => {
             </p>
           </WarningCallout>
         )}
-        {FormRUtilities.displaySubmissionDate(
-          formData.submissionDate,
-          "submissionDateTop"
-        )}
+        {!canEdit &&
+          FormRUtilities.displaySubmissionDate(
+            formData.submissionDate,
+            "submissionDateTop"
+          )}
         <ViewSection1 {...viewSectionProps} />
         <ViewSection2 {...viewSectionProps} />
         <ViewSection3 {...viewSectionProps} />
@@ -83,10 +84,11 @@ const View = ({ canEdit, history }: IView) => {
         <ViewSection6 {...viewSectionProps} />
         {enableCovidDeclaration ? <ViewSection7 {...viewSectionProps} /> : null}
         {!canEdit && <ViewSection8 {...viewSectionProps} />}
-        {FormRUtilities.displaySubmissionDate(
-          formData.submissionDate,
-          "submissionDate"
-        )}
+        {!canEdit &&
+          FormRUtilities.displaySubmissionDate(
+            formData.submissionDate,
+            "submissionDate"
+          )}
       </>
     );
   else content = <Redirect to="/formr-b" />;
