@@ -1,17 +1,22 @@
 import PersonalDetailsComponent from "./personal-details/PersonalDetailsComponent";
-import Programmes from "./programmes/Programmes";
-import Placements from "./placements/Placements";
 import { Fieldset, Details } from "nhsuk-react-components";
 import PageTitle from "../common/PageTitle";
 import ScrollTo from "../forms/ScrollTo";
-
-import { useAppSelector } from "../../redux/hooks/hooks";
-import { selectTraineeProfile } from "../../redux/slices/traineeProfileSlice";
 import DataSourceMsg from "../common/DataSourceMsg";
 import { Redirect } from "react-router-dom";
+import { PanelsCreator } from "./PanelsCreator";
+import { ProfileType, TraineeProfileName } from "../../models/TraineeProfile";
+import style from "../Common.module.scss";
+import { PANEL_KEYS } from "../../utilities/Constants";
+import { useAppSelector } from "../../redux/hooks/hooks";
+import { selectTraineeProfile } from "../../redux/slices/traineeProfileSlice";
+import { placementPanelTemplate } from "../../models/Placement";
+import { programmePanelTemplate } from "../../models/ProgrammeMembership";
 
 const Profile = ({ mfa }: any) => {
-  const traineeProfileData = useAppSelector(selectTraineeProfile);
+  const placementsArr = useAppSelector(selectTraineeProfile).placements;
+  const programmesArr =
+    useAppSelector(selectTraineeProfile).programmeMemberships;
 
   if (mfa === "NOMFA") {
     return <Redirect to="/mfa" />;
@@ -23,7 +28,7 @@ const Profile = ({ mfa }: any) => {
       <Fieldset>
         <Fieldset.Legend
           isPageHeading
-          style={{ color: "#005EB8" }}
+          className={style.fieldLegHeader}
           data-cy="profileHeading"
         >
           Profile
@@ -31,19 +36,25 @@ const Profile = ({ mfa }: any) => {
       </Fieldset>
       <DataSourceMsg />
       <Details.ExpanderGroup>
-        {traineeProfileData.personalDetails && (
-          <PersonalDetailsComponent
-            personalDetails={traineeProfileData.personalDetails}
-          />
-        )}
-        {traineeProfileData.placements && (
-          <Placements placements={traineeProfileData.placements}></Placements>
-        )}
-        {traineeProfileData.programmeMemberships && (
-          <Programmes
-            programmeMemberships={traineeProfileData.programmeMemberships}
-          ></Programmes>
-        )}
+        <PersonalDetailsComponent />
+        <PanelsCreator
+          panelsArr={prepareProfilePanelsData(
+            placementsArr,
+            TraineeProfileName.Placements
+          )}
+          panelsName={TraineeProfileName.Placements}
+          panelsTitle={PANEL_KEYS.placements}
+          panelKeys={PANEL_KEYS}
+        />
+        <PanelsCreator
+          panelsArr={prepareProfilePanelsData(
+            programmesArr,
+            TraineeProfileName.Programmes
+          )}
+          panelsName={TraineeProfileName.Programmes}
+          panelsTitle={PANEL_KEYS.programmeMemberships}
+          panelKeys={PANEL_KEYS}
+        />
       </Details.ExpanderGroup>
     </div>
   );
@@ -51,3 +62,32 @@ const Profile = ({ mfa }: any) => {
 };
 
 export default Profile;
+
+function prepareProfilePanelsData(
+  arr: ProfileType[],
+  arrName: TraineeProfileName
+) {
+  return arr.map((obj, _index) => filterAndOrderProfilePanelData(arrName, obj));
+}
+
+function filterAndOrderProfilePanelData<T>(
+  pName: TraineeProfileName,
+  pObj: T extends ProfileType ? any : any
+) {
+  if (pName === TraineeProfileName.Placements) {
+    const reorderedPl = Object.assign(placementPanelTemplate, { ...pObj });
+    const { tisId, status, ...filteredPlacementPanel } = reorderedPl;
+    return filteredPlacementPanel;
+  } else {
+    const reorderedPr = Object.assign(programmePanelTemplate, { ...pObj });
+    const {
+      tisId,
+      programmeTisId,
+      programmeMembershipType,
+      status,
+      programmeCompletionDate,
+      ...filteredProgrammePanel
+    } = reorderedPr;
+    return filteredProgrammePanel;
+  }
+}
