@@ -117,7 +117,7 @@ export const draftFormRPartBWithNullCareerBreak: FormRPartB = {
   ],
   sicknessAbsence: 0,
   parentalLeave: 0,
-  careerBreaks: null,
+  careerBreaks: 0,
   paidLeave: 0,
   unauthorisedLeave: 10,
   otherLeave: 0,
@@ -195,7 +195,7 @@ export const draftFormRPartBWithNoLeaveTotal = {
   paidLeave: 1,
   unauthorisedLeave: 1,
   otherLeave: 1,
-  totalLeave: null,
+  totalLeave: 0,
   isHonest: true,
   isHealthy: true,
   isWarned: true,
