@@ -4,11 +4,17 @@ import { PersonalDetails } from "./PersonalDetails";
 
 export interface TraineeProfile {
   traineeTisId: string;
-  personalDetails: PersonalDetails | null;
+  personalDetails: PersonalDetails;
   programmeMemberships: ProgrammeMembership[];
   placements: Placement[];
 }
 
+export type ProfileType = Placement | ProgrammeMembership;
+
+export enum TraineeProfileName {
+  Placements = "placements",
+  Programmes = "programmeMemberships"
+}
 export enum PanelName {
   Placement = "placement",
   Programme = "programme"
