@@ -3,6 +3,7 @@ import { AxiosResponse } from "axios";
 import { TraineeProfile } from "../../models/TraineeProfile";
 import { TraineeProfileService } from "../../services/TraineeProfileService";
 import { initialPersonalDetails } from "../../models/PersonalDetails";
+import { DateUtilities } from "../../utilities/DateUtilities";
 
 interface IProfile {
   traineeProfileData: TraineeProfile;
@@ -33,7 +34,6 @@ export const fetchTraineeProfileData = createAsyncThunk(
   }
 );
 
-// TODO CORS error if addCredential made via localhost
 export const issueDspCredential = createAsyncThunk(
   "traineeProfile/issueDspCredential",
   async (parData: { panelId: string; panelName: string }) => {
@@ -43,7 +43,9 @@ export const issueDspCredential = createAsyncThunk(
       panelId,
       panelName
     );
-    window.location.href = response.headers.location;
+    if (!!response.headers.location) {
+      window.location.href = response.headers.location;
+    }
   }
 );
 
@@ -68,7 +70,21 @@ const traineeProfileSlice = createSlice({
       })
       .addCase(fetchTraineeProfileData.fulfilled, (state, action) => {
         state.status = "succeeded";
-        state.traineeProfileData = action.payload;
+        state.traineeProfileData.traineeTisId = action.payload.traineeTisId;
+        state.traineeProfileData.personalDetails =
+          action.payload.personalDetails;
+        const sortedProgrammes = DateUtilities.genericSort(
+          action.payload.programmeMemberships,
+          "startDate",
+          true
+        );
+        state.traineeProfileData.programmeMemberships = sortedProgrammes;
+        const sortedPlacements = DateUtilities.genericSort(
+          action.payload.placements,
+          "startDate",
+          true
+        );
+        state.traineeProfileData.placements = sortedPlacements;
       })
       .addCase(fetchTraineeProfileData.rejected, (state, action) => {
         state.status = "failed";
